Fixes #76 by moving the contents of adjust_compiler into the build_ext command itself.  This way it doesn't run unless an extension module actually has to be compiled.  Looking back to the issues where the adjust_compiler was originally added, I don't think it really needs to happen any sooner than it does here.  The original adjust_compiler function becomes a no-op and is deprecated.  It can be removed from any setup.py that uses it (and has the updated astropy_helpers)
